입력데이어 유효성체크

+++ src/main/java/com/hyo/untactStudent/controller/UsrArticleController.java
@@ -63,12 +63,18 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(Integer id) {
+		
+		if( id == null) {
+			return new ResultData("F-1", "번호를 입력해주세요");
+		}
+		
 		Article article =articleService.getArticle(id);
 		
 		if( article == null) {
 			return new ResultData( "F-1", "없는 게시물입니다.");
 		}
+	
 		
 		
 		return articleService.deleteArticle(id);
@@ -81,7 +87,20 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body) {
+		if( id == null) {
+			return new ResultData("F-1", "번호를 입력해주세요");
+		}
+		
+		if( title == null) {
+			return new ResultData("F-1", "제목을 입력해주세요");
+
+		}
+		if( body == null) {
+			return new ResultData("F-1", "내용을 입력해주세요");
+
+		}
+		
 		Article article =articleService.getArticle(id);
 		if( article == null) {
 			return new ResultData( "F-1", "없는 게시물입니다.");
